+++ packages/components/color-picker/props.ts
@@ -11,10 +11,20 @@ const props: TdColorPickerProps = {
     type: Boolean,
     value: true,
   },
+  /** 颜色模式选择。同时支持单色和渐变两种模式，可仅使用单色或者渐变其中一种模式，也可以同时使用。`monochrome` 表示单色，`linear-gradient` 表示渐变色 */
+  colorModes: {
+    type: null,
+    value: 'monochrome',
+  },
   /** 是否开启透明通道 */
   enableAlpha: {
     type: Boolean,
     value: false,
+  },
+  /** 是否允许开启通过点击渐变轴增加渐变梯度，默认开启，关闭时只会存在起始和结束两个颜色 */
+  enableMultipleGradient: {
+    type: Boolean,
+    value: true,
   },
   /** 如果 color-picker 是在一个 `position:fixed` 的区域，需要显式指定属性 fixed 为 true */
   fixed: {