+++ packages/pro-components/chat/_example/base/mock.js
@@ -1,7 +1,0 @@
-export function getReasoning() {
-  return '1';
-}
-
-export function getReasoning2() {
-  return '2';
-}