+++ packages/pro-components/chat/chat-message/_example/configure/index.js
@@ -1,0 +1,13 @@
+Component({
+  data: {
+    message: {
+      role: 'system',
+      content: [
+        {
+          type: 'text',
+          data: '牛顿第一定律是否适用于所有参考系？',
+        },
+      ],
+    },
+  },
+});