+++ script/gulpfile.example.js
@@ -18,7 +18,7 @@ const dist = '_example';
 const { clear, build: baseBuild, watch: baseWatch, handleError, resetError } = base(src, dist, 'example');
 
 // src component examples
-const srcExampleInput = ['packages/components/**/_example/**/*.*'];
+const srcExampleInput = ['packages/components/**/_example/**/*.*', 'packages/pro-components/chat/**/_example/**/*.*'];
 const srcExampleOutput = 'packages/tdesign-miniprogram/example/pages';
 const copySrcExample = () => {
   return gulp